Page->Window 유형 변경
MainWindow에서 Page를 띄우지 않고,
MainWindow를 닫고, ChatWindow를 여는 방식으로 변경.
