--- Client/ChatPage.xaml
@@ -1,33 +1,0 @@
-﻿<Page x:Class="Client.ChatPage"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:Client"
-      PreviewKeyDown="Window_PreviewKeyDown"
-      mc:Ignorable="d" 
-      d:DesignHeight="500" d:DesignWidth="850"
-      Title="ChatPage">
-
-    <Grid>
-        <ListView x:Name="UserListBox" d:ItemsSource="{d:SampleData ItemCount=20}" Margin="646,112,0,0">
-            <ListView.View>
-                <GridView>
-                    <GridViewColumn/>
-                </GridView>
-            </ListView.View>
-        </ListView>
-        <TextBox x:Name="ChatBox" HorizontalAlignment="Left" Margin="0,429,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="584" Height="76"/>
-        <Button x:Name="SendButton" Content="Send" HorizontalAlignment="Left" Margin="584,429,0,0" VerticalAlignment="Top" Height="76" Width="62" Click="OnSendButton"/>
-        <TextBlock x:Name="NickNameBlock" HorizontalAlignment="Left" Margin="665,10,0,0" Foreground="Green" TextWrapping="Wrap" Text="TextBlock" FontSize="20" VerticalAlignment="Top" Height="27" Width="117"/>
-        <Button x:Name="SignOutButton" Content="Sign Out" HorizontalAlignment="Left" Margin="778,75,0,0" Click="OnSignOutButton" VerticalAlignment="Top" Width="62" Height="32"/>
-        <ListView x:Name="ChatListView" d:ItemsSource="{d:SampleData ItemCount=5}" Margin="0,0,204,71">
-            <ListView.View>
-                <GridView>
-                    <GridViewColumn/>
-                </GridView>
-            </ListView.View>
-        </ListView>
-
-    </Grid>
-</Page>
