+++ Client/ChatWindow.xaml
@@ -1,0 +1,42 @@
+﻿<Window x:Class="Client.ChatWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:Client"
+        mc:Ignorable="d"
+        PreviewKeyDown="Window_PreviewKeyDown"
+        Title="ChatWindow" Height="520" Width="910" ResizeMode="NoResize">
+    <Grid>
+        <ListView x:Name="UserListBox" d:ItemsSource="{d:SampleData ItemCount=20}" Margin="694,112,0,16"
+                  VerticalContentAlignment="Top"
+                  ScrollViewer.VerticalScrollBarVisibility="Visible"
+                  ScrollViewer.CanContentScroll="False"
+                  ScrollViewer.HorizontalScrollBarVisibility="Disabled">
+            <ListView.View>
+                <GridView>
+                    <GridViewColumn Header="Current User List" Width="100"/>
+                </GridView>
+            </ListView.View>
+        </ListView>
+
+        <TextBox x:Name="ChatBox" HorizontalAlignment="Left" Margin="0,422,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="632" Height="66"/>
+        <Button x:Name="SendButton" Content="Send" HorizontalAlignment="Left" Margin="632,422,0,0" VerticalAlignment="Top" Height="66" Width="62" Click="OnSendButton"/>
+        <TextBlock x:Name="NickNameBlock" HorizontalAlignment="Left" Margin="710,15,0,0" Foreground="Green" TextWrapping="Wrap" Text="TextBlock" FontSize="20" VerticalAlignment="Top" Height="60" Width="184"/>
+        <Button x:Name="SignOutButton" Content="Sign Out" HorizontalAlignment="Left" Margin="819,75,0,0" Click="OnSignOutButton" VerticalAlignment="Top" Width="62" Height="32" />
+
+        <ListView x:Name="ChatListView" d:ItemsSource="{d:SampleData ItemCount=5}" Margin="0,0,216,82"
+                  VerticalContentAlignment="Bottom"
+                  ScrollViewer.VerticalScrollBarVisibility="Visible"
+                  ScrollViewer.CanContentScroll="False"
+                  ScrollViewer.HorizontalScrollBarVisibility="Disabled">
+
+            <ListView.View>
+                <GridView>
+                    <GridViewColumn Header="ID" Width="100"/>
+                    <GridViewColumn Header="Content" Width="500"/>
+                </GridView>
+            </ListView.View>
+        </ListView>
+    </Grid>
+</Window>
